// client/src/types/ADSB_Type.ts
export type Point = [number,number]

export type Category = null | "No ADS-B Emitter Category Information" | "Light" | "Small" | "Large" | "High Vortex Large" | "Heavy" | "High Performance" | "Rotorcraft" | "Glider, sailplane" | "Lighter-than-air" | "Parachutist / Skydiver" | "Ultralight, hang-glider, paraglider" | "Reserved" | "Unmanned Aerial Vehicle" | "Space, Trans-atmospheric vehicle" | "Surface Vehicle - Emergency Vehicle" | "Surface Vehicle - Service Vehicle" | "Point Obstacle" | "Cluster Obstacle" | "Line Obstacle"



// ADSB generic state
export interface ADSB_State {
    "time": number, // When the state was recieved as epoch in milliseconds
    "states": Aircraft_State[], // List of aircraft states
    "source": string // name of service that data was from
}

// ADS-B information at a given time point
export interface Aircraft_State {
    "icao24":string, // ADS-B ICAO24
    "callsign":string|null, // Callsign
    "squawk":string|null, // Squawk
    "emergency":string|null, // Emergency
    "spi": number|null, // SPI (special purpose indicator)
    "time":number|null, // Time position was sent
    "lon":number|null, // Longitude
    "lat":number|null, // Latitude
    "alt":number|null, // Altitude (geometric preferred but may be baro)
    "heading":number|null, // Heading, 0 is north
    "roll":number|null, // Roll amount
    "velocity":number|null, // Velocity in MPH (ground > true > indicator)
    "vertical_rate":number|null, // Vertical rate in FPS
    "on_ground":boolean|null, // On ground or not
    "category":Category // Aircraft type
}